Convert generations to int

+++ exprimo/utils/parameter_search.py
@@ -45,7 +45,7 @@ def do_parameter_search(config_path, parameter_grid, repeats=10, verbose=False):
         args = dict(tuple(args.items()) + combination)
 
         if 'generations' in args:
-            args['generations'] = args['generations'] * (original_pop / args['population_size'])
+            args['generations'] = int(args['generations'] * (original_pop / args['population_size']))
 
         current_config = config.copy()
         current_config['optimizer_args'] = args
